Software/server/requirements.txt
requests
paho-mqtt
numpy
torch
matplotlib
pymongo
flask
asyncio
aiohttp
pydantic